feat: migrate `assocPath` to a separate module
