+++ .changeset/fresh-fishes-dream.md
@@ -1,0 +1,6 @@
+---
+"astro-reading-time": patch
+"@mogeko/utils": patch
+---
+
+Migrate `assocPath` to a separate module.
