+++ packages/utils/README.md
@@ -1,0 +1,3 @@
+# @mogeko/utils
+
+Useful function snippets for Node.js.
